work in progress
Former-commit-id: 326e640a72eb9131d0a3aa6bfb2ea90502f03e30
Former-commit-id: d7d3b04bba12cf7590e38ca9d3ffc73a301edbdb

+++ src/thing.h
@@ -1109,7 +1109,7 @@ static inline uint8_t thing_is_fire (thingp t)
     return (tp_is_fire(thing_tp(t)));
 }
 
-static inline uint8_t thing_collision_radius (thingp t)
+static inline double thing_collision_radius (thingp t)
 {
     verify(t);
